Semana7/userManagement/src/modules/auth/application/__init__.py
from pydispatch import dispatcher
from .handlers import DomainCredentialHandler

#dispatcher.connect(DomainCredentialHandler.handle_credential_created, signal='CredentialCreatedDomain')
#dispatcher.connect(DomainCredentialHandler.handle_credential_created, signal='CredentialCreatedIntegration')
