Semana7/userManagement/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
flask-restful==0.3.9
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
psycopg2-binary==2.9.6
PyJWT==2.1.0
requests==2.26.0
SQLAlchemy==1.4.15
Werkzeug==2.2.2
PyDispatcher==2.0.6
pulsar-client[avro]==3.4.0
